pyproject.toml
[tool.poetry]
name = "plutus"
version = "0.1.0"
description = "Financial Portfolio management tools"
authors = ["madhukar01 <mholla8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyMuPDF = "^1.18.15"
camelot-py = "^0.10.1"
opencv-python = "^4.5.3"
ghostscript = "^0.7"

[tool.poetry.dev-dependencies]
toml = "^0.10.2"
yapf = "^0.31.0"
isort = "^5.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
default_section = "THIRDPARTY"
known_first_party = ["intel_app_platform", "hsutils", "utils"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
lines_between_sections = 1
force_grid_wrap = 0
use_parentheses = true
multi_line_output = 2
float_to_top = true

[tool.yapf]
based_on_style = "google"
column_limit = 79
